Deduplicate user storage test helpers
